backend/pyproject.toml
# @spec: specs/002-fullstack-web-app/plan.md
# @spec: specs/003-ai-chatbot/research.md (Technology Stack Summary)
# Backend project configuration for FastAPI Todo application with AI chatbot

[project]
name = "todo-backend"
version = "0.2.0"
description = "FastAPI backend for Todo application with AI chatbot"
requires-python = ">=3.13"
dependencies = [
    # Phase II dependencies
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.0",
    "pytest>=8.3.0",
    "httpx>=0.28.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "pyjwt>=2.10.1",
    "bcrypt>=5.0.0",
    # Database migrations
    "alembic>=1.14.0",
    # Phase III dependencies (AI chatbot)
    # OpenAI Agents SDK (installed from official GitHub repo)
    "openai-agents>=0.1.0",
    "openai>=1.60.0",
    "slowapi>=0.1.9",
    "aiosqlite>=0.20.0",  # For async SQLite support
    "asyncpg>=0.31.0",
    # Official MCP SDK (FastMCP) for Model Context Protocol
    "fastmcp>=0.1.0",
    "mcp>=0.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
